workdir: Remove make-hierarchy script from each app directory
`make-hierarchy` is run for each application directory only at the
beginning (as part for `fill-workdir.sh`). So there is no need to have
it copied in each application directory. Move it to
`copy/common-nocopy/` and run it from there.

If users require running it from an application directory, they could
run it from `../../copy/common-nocopy/`.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>