Refactor app build and run scripts
Currently, build and run scripts are copied for each type of build
required. This is difficult to maintain: there are multiple applications
with multiple scripts. Adding a new application or a new configuration
requires creating new build and run scripts.

Refactor this by creating a single `build.sh` script and a single
`run.sh`. These are then symlinked into each application directory, with
a unique symbolic link name that follows the conventions for config
files:

.config.<appname>-<plat>-<arch>[-<fs>][-<extra>]

clang-build-<plat>-<arch>[-<fs>][-<extra>].sh
gcc-build-<plat>-<arch>[-<fs>][-<extra>].sh
run-<plat>-<arch>[-<fs>].sh

Inside the `build.sh` and `run.sh` scripts, the script name (now the
link name) is checked (the `$0` variable). Based on the name,
appropriate actions happen: build with GCC for QEMU on AArch64, build
with Clang for Firecracker on x86_64 with initrd filesystem and PIE, run
QEMU on x86_64 with 9pfs support and networking.

The `build.sh` and `run.sh` scripts are stored in the `copy/link`
directory. In order to create symbolic links to the application
directory, the `copy/common-nocopy/generate.sh` script is used. The
`generate.sh` script is invoked at the setup phase (inside the
`fill-workdir.sh` script). It walks through config files in the
application directory and creates corresponding symbolic links to the
`build.sh` and `run.sh` scripts.

It is expected that properly named config files are part of the
application directory. That is, the configuration files (used by `make
defconfig`) follow the above convention:

.config.<appname>-<plat>-<arch>[-<fs>][-<extra>]

Where config files are not part of the actual application repository,
they will be stored in the `copy/base/apps/<appname>` directory and they
will be copied in the application directory at the setup phase.

An example of a config file name, together with the corresponding build
and run scripts (symbolic links to `build.sh` and `run.sh`):

.config.helloworld-qemu-x86_64-pie
gcc-build-qemu-x86_64-pie.sh
clang-build-qemu-x86_64-pie.sh
run-qemu-x86_64.sh

After building a unikernel image, a copy of the final image is created
in the application directory to save it for further runs (without a
build required).

When running a unikernel, certain files may be part of the application
directory (these are to be copied from `copy/base/apps/<appname>`):

* `network`: If this file is present (no need for contents), networking
  will be used to run the application.
* `kvm`: If this file is present (no need for contents), hardware
  acceleration (KVM) will be used with QEMU to run the application.
* `extra_boot_args`: If this file is present, its contents will be
  appended to each unikernel boot comand. This is useful for
  running applications such as Redis, that require the configuration
  file to be passed as a boot time argument.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>