Add testing scripts
Add a script that will go through all the applications, build them with
all available compiler and config options combinations, and runs the
application for every build.

```
./test-apps.sh

Testing builds for apps/helloworld/:
building with ./clang-11-build-fc-x86_64-pie.sh		PASSED
Running apps/helloworld/ on fc x86_64 			PASSED
building with ./clang-11-build-fc-x86_64.sh		PASSED
Running apps/helloworld/ on fc x86_64 			PASSED
[...]
```

Logs of both builds and runs are saved in `logs/builds/app-name/` and
`logs/runs/app-name`.

Beware that it takes a lot of time for the script to run, since there
are a lot of different build to be done for every application.

In order for the script to be fully functionable, config files need to
be added (under `workdir/copy/base/apps/app-name/`) for every
application. This is because if no config file is present, the build
scripts will open a menuconfig screen and hang.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>